Create listOfPages.java
linked list of  pages.  Forget how to work pointers in java and I didn't have wifi when I wrote this please forgive me I'll change that later.  Made lots of changes here allow me to explain.  This is a doubly linked list of nodes, each containing a panel.  Right now its 25 nodes long but we can change that to fit however many pages we have obviously.  Instead of having a tutor object in universe, I'm going to have a linkedlist object which has functions that call different constructors of the Tutor class.  These constructors should create panels that display a video, pdf, html, or a quiz question.  We can tweak that later those are just the four that came to mind.  This way it will be easy to add new pages to our project all we have to do is add a node to the array and call the function passing the item we would like to display.